+++ src/cpx_io/cpx_system/cpx_ap/ap_supported_functions.py
@@ -38,7 +38,6 @@ SUPPORTED_PRODUCT_FUNCTIONS_DICT = {
     "read_output_channel": [
         ProductCategory.ANALOG,
         ProductCategory.DIGITAL,
-        ProductCategory.IO_LINK,
         ProductCategory.VTOM,
         ProductCategory.VTSA,
         ProductCategory.VTUG,